+++ src/Polaroider/SnapshotOptions.cs
@@ -299,7 +299,19 @@ namespace Polaroider
 		}
 
 		/// <summary>
-		/// set an expression to evaluate valuetypes. defaults to (type, obj) => (type.IsValueType || type == typeof(string)) && !type.IsGenericType
+		/// Ignore whitespaces in the snapshot
+		/// </summary>
+		/// <param name="options"></param>
+		/// <returns></returns>
+        public static SnapshotOptions IgnoreWhiteSpaces(this SnapshotOptions options)
+        {
+			options.AddDirective(line => line.Replace(" ", string.Empty));
+
+			return options;
+        }
+
+		/// <summary>
+		/// set an expression to evaluate valuetypes. defaults to (type, obj) => (type.IsValueType || type == typeof(string)) &amp;&amp; !type.IsGenericType
 		/// </summary>
 		/// <param name="options"></param>
 		/// <param name="evaluator"></param>