+++ src/Polaroider/SnapshotReaders/SettingsReader.cs
@@ -1,15 +1,26 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
+﻿
 namespace Polaroider.SnapshotReaders
 {
+    /// <summary>
+    /// 
+    /// </summary>
     internal class SettingsReader : ILineReader
     {
-        public void ReadLine(string line, Snapshot item)
+        /// <summary>
+        /// Read the line
+        /// </summary>
+        /// <param name="line"></param>
+        /// <param name="snapshot"></param>
+        public void ReadLine(string line, Snapshot snapshot)
         {
+            // do nothing for settings
         }
 
+        /// <summary>
+        /// Defaults to false
+        /// </summary>
+        /// <param name="snapshot"></param>
+        /// <returns></returns>
         public bool NewSnapshot(Snapshot snapshot)
         {
             return false;