+++ src/Polaroider/SnapshotReaders/SnapshotLineReader.cs
@@ -1,15 +1,28 @@
 ﻿
 namespace Polaroider.SnapshotReaders
 {
+    /// <summary>
+    /// 
+    /// </summary>
     internal class SnapshotLineReader : ILineReader
     {
-        public void ReadLine(string line, Snapshot item)
+        /// <summary>
+        /// Read the line and add it to the snapshot
+        /// </summary>
+        /// <param name="line"></param>
+        /// <param name="snapshot"></param>
+        public void ReadLine(string line, Snapshot snapshot)
         {
             // set data row
             var row = new Line(line);
-            item.Add(row);
+            snapshot.Add(row);
         }
         
+        /// <summary>
+        /// Defaults to false
+        /// </summary>
+        /// <param name="snapshot"></param>
+        /// <returns></returns>
         public bool NewSnapshot(Snapshot snapshot)
         {
             return false;