+++ src/Polaroider/SnapshotSetupResolver.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Reflection;
+using System.Text;
 
 namespace Polaroider
 {
@@ -46,9 +47,9 @@ namespace Polaroider
 
                 if (string.IsNullOrEmpty(name))
                 {
-                    // stackFrame.GetFileName() only works when pdb files are provided
+                    // stackFrame.GetFileName() only works when pdb files are provided and the code is not optimized
                     // When using Live Unit Testing the checkbox for "Enable debug symbol..." has to be activated
-                    throw new InvalidOperationException($"Polaroider could not find the file containing the Test {method.Name}.{Environment.NewLine}Polaroider relies on pdb to determine the path of the executing testclass.{Environment.NewLine}Please ensure that pdb files are generated when building the projects.{Environment.NewLine}When using Live Unit Testing make sure the checkbox for 'Enable debug symbol and xml documentation comment generation' is enabled in the Visual Studio Options.");
+                    throw new TestMethodNotFoundException();
                 }
 
                 return new SnapshotSetup(name, method);
@@ -59,7 +60,8 @@ namespace Polaroider
                 return new SnapshotSetup(lastFrame.GetFileName(), lastFrame.GetMethod());
             }
 
-            return null;
+            // stackFrame.GetFileName() only works when pdb files are provided and the code is not optimized
+            throw new TestMethodNotFoundException();
         }
 
         private bool IsInternalMethod(MemberInfo method)
@@ -83,9 +85,6 @@ namespace Polaroider
             {
                 Func<string, bool> methodContainsAttribute = attributeName =>
                 {
-                    //var attribute = method?.CustomAttributes.FirstOrDefault(a =>
-                    //{
-                    //    var type = a.AttributeType;
                     var attribute = method?.GetCustomAttributes()?.FirstOrDefault(a =>
                     {
                         var type = a.GetType();