+++ docs/changelog.md
@@ -5,8 +5,14 @@ nav_order: 99
 ---
 
 ## Polaroider Changelog
-### vNext
-
+### v2.0.7
+#### Added
+- Added ExtensionMethod for IgnoreWhiteSaces
+  
+#### Fixed
+- NullReferenceException when using optimized compiling
+- Documentation on Optimize code option
+  
 ### v2.0.6
 - Addd nuke for building
 - Troubleshooting when using Live Unit Testing